Refactors form
