+++ client/src/components/DeliveryInformation/Form/Form.module.css
@@ -1,0 +1,3 @@
+.container {
+  padding: 1rem;
+}
